requirements.txt
python-telegram-bot==21.0.1
youtube-transcript-api==0.6.2
google-api-python-client==2.111.0
anthropic==0.39.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
pytest-cov==4.1.0
